requirements.txt
streamlit
pandas
numpy
spicy
sqlalchemy
pymysql
requests
python-dotenv